move 'selected-*' functions to query_function.py
