+++ corehq/apps/case_search/xpath_functions/value_functions.py
@@ -6,7 +6,6 @@ from django.utils.translation import ugettext as _
 from eulxml.xpath.ast import serialize
 
 from corehq.apps.case_search.exceptions import XPathFunctionException
-from corehq.apps.es.case_search import case_property_query
 
 
 def date(node):
@@ -39,22 +38,3 @@ def date(node):
         "The \"date\" function only accepts integers or strings of the format \"YYYY-mm-dd\"",
         serialize(node)
     )
-
-
-def selected_any(domain, node, fuzzy):
-    return _selected_query(node, fuzzy=fuzzy, operator='or')
-
-
-def selected_all(domain, node, fuzzy):
-    return _selected_query(node, fuzzy=fuzzy, operator='and')
-
-
-def _selected_query(node, fuzzy, operator):
-    if len(node.args) != 2:
-        raise XPathFunctionException(
-            _("The {name} function accepts exactly two arguments.").format(name=node.name),
-            serialize(node)
-        )
-    property_name = serialize(node.args[0])
-    search_values = node.args[1]
-    return case_property_query(property_name, search_values, fuzzy=fuzzy, multivalue_mode=operator)
